Clean up `serenity::model::prelude`
