+++ src/model/channel/message.rs
@@ -511,7 +511,7 @@ impl Message {
     /// does not have the required [permissions].
     ///
     /// [Add Reactions]: Permissions::ADD_REACTIONS
-    /// [permissions]: super::permissions
+    /// [permissions]: crate::model::permissions
     #[inline]
     pub async fn react(
         &self,
